+++ substratevm/src/com.oracle.svm.hosted/src/com/oracle/svm/hosted/code/HostedReplacements.java
@@ -75,7 +75,7 @@ public class HostedReplacements extends SubstrateReplacements {
     @Override
     public void registerSnippet(ResolvedJavaMethod method, ResolvedJavaMethod original, Object receiver, boolean trackNodeSourcePosition, OptionValues options) {
         /* We must have the snippet already available in the analysis replacements. */
-        assert aReplacements.getSnippet(((HostedMethod) method).wrapped, null, null, trackNodeSourcePosition, null, options) != null;
+        assert aReplacements.getSnippet(((HostedMethod) method).wrapped, null, null, null, trackNodeSourcePosition, null, options) != null;
     }
 
     @Override